// frontend/src/pages/statisticReport/models/Statistic.ts
export interface Statistic {
    id: number;
    creationDate: Date;
    totalLandSurveys: number;
    averageValue: number;
    totalCarles: number;
    totalAgencies: number;
    totalOwners: number;
    totalCarlesRed: number;
    totalReds: number;
    totalNormals: number;
    totalGreens: number;
    belowMinLimit: number;
    belowMaxLimit: number;
    overLimits: number;
    WithTitle: number;
    totalUnworkable: number;
    posadasTotal: number;
    garupaTotal: number;
    candelariaTotal: number;
}